[UPD] Refactor base firmware files
